Block Editor: Replace dependency on Core Data with an Editor filter.
