+++ packages/editor/src/hooks/index.js
@@ -1,4 +1,5 @@
 /**
  * Internal dependencies
  */
+import './custom-sources-backwards-compatibility';
 import './default-autocompleters';
